* Change old cast to new cast for some part of code * Change NULL to nullptr for some part of code * Remove pre processor #ifndef in some .hpp
* Add export feature in GUI with help of new class DM_ResultGroupGui and DM_ItemGroupGui
* Display input model in bold when obligatory and in italic when optionnal in GAboutStepDialog
* Correct a bug in GMainWindow when restore maximized state (the window become smaller at each start of application) 
* Correct a bug in CG_CustomTreeItemModel that can cause a crash when remove elements from tree view
* Correct a bug in treeview that doesn't display the displayable name of the model in some situation
* Correct a bug in CDM_StepManager when a step want to restart another step
* Correct a bug in opf reader that don't create models in same order each time a same file is readed.
* Add autodelete feature for CT_AttributesColor, CT_AttributesNormal and CT_AttributesScalarT
* Add NodeEditor in ctlibmodelsextraview and change some part of code to display a node graph when the user want to link input model with output model instead of the old tree view of commputree v5
* Restore metrics and filters capability
* Add macros CT_TYPE_IMPL_INIT_MACRO and CT_TYPE_TEMPLATED_IMPL_OPENCV_INIT_MACRO and CT_TYPE_TEMPLATED_IMPL_VALUES_INIT_MACRO to correct a problem in computree v5 that does'nt display name of type if the class is not instancied at least one time at startup.
* Correct method shortDescription and detailledDescription of CT_AbstractModel to display at least the displayable name if the description is empty
* Correct a bug in CT_InResultModelGroup and CT_InResultModelGroupToCopy that disable the creation of an optionnal or recursive result
* Correct a bug in CT_InStdSingularItemModel when it was copied (problem with attributes)
* Add methods in CT_HandleInBase to get some extra models
* Correct a bug in CT_InResultModelConfigurationManager that cause a problem with readOnly view
* Implement the method selectOneGraphForPossibilityOfRootModel in CT_ModelSelectionHelper to select a default graph before user see the node editor view
* Correct a bug in CT_ContainerT that cause crash in some situation when copy a result
* Correct a bug in CT_TNodeGroup and CT_TTreeGroup that cause crash when copy a result
* Correct a bug in CT_ReaderItem when copy it
* Correct a bug in CT_VirtualAbstractStep that don't call finalizeXXXSettings in restorePreSettings and restorePostSettings
* Refactor begin and end loop step
* Correct a bug in ct_view/CT_XXX.cpp that don't restore settings in attribute of the step but only in view !

git-svn-id: http://rdinnovation.onf.fr/svn/computreev6@48 b135c390-378b-4ba5-bdcd-e4540112eb5a

+++ library/ctlibstructureaddon/ct_itemdrawable/ct_loopcounter.h
@@ -48,14 +48,14 @@ public:
     CT_LoopCounter(QSharedPointer<CT_Counter> counter, CT_StepBeginLoop* beginStep);
     CT_LoopCounter(const CT_LoopCounter& other) = default;
 
-    inline size_t nTurns() const {return _counter->nTurns();}
-    inline size_t currentTurn() const {return _counter->currentTurn();}
+    inline int nTurns() const {return _counter->nTurns();}
+    inline int currentTurn() const {return _counter->currentTurn();}
     inline QString turnName() const {return _counter->turnName();}
     inline bool hasNextTurn() {return _counter->hasNextTurn();}
 
     inline void beginNextTurn() {_counter->beginNextTurn();}
 
-    inline void setCurrentTurn(size_t n) {_counter->setCurrentTurn(n);}
+    inline void setCurrentTurn(int n) {_counter->setCurrentTurn(n);}
 
     inline void setTurnName(QString name) {_counter->setTurnName(name);}
