+++ library/ctlibstructureaddon/ct_itemdrawable/tools/ct_counter.cpp
@@ -1,6 +1,6 @@
 #include "ct_counter.h"
 
-CT_Counter::CT_Counter(const size_t &nTurns)
+CT_Counter::CT_Counter(const int& nTurns)
 {
     _nTurns = nTurns;
     _currentTurn = 1;
